+++ src/expression/rawparameter.cpp
@@ -16,7 +16,7 @@ void rawparameter::synchronize(void)
     opnums = std::vector<int>(universe::mymesh->getdisjointregions()->count(),-1);
 
     // Rebuild the structure:
-    for (int i = 0; i < mystructuretracker.size(); i++)
+    for (size_t i = 0; i < mystructuretracker.size(); i++)
         set(mystructuretracker[i].first, mystructuretracker[i].second);
     
     
@@ -28,7 +28,7 @@ void rawparameter::errorifundefined(std::vector<int> disjregs)
 {
     synchronize();
 
-    for (int i = 0; i < disjregs.size(); i++)
+    for (size_t i = 0; i < disjregs.size(); i++)
     {
         if (myoperations[disjregs[i]].size() == 1 && myoperations[disjregs[i]][0] == NULL)
         {
@@ -43,7 +43,7 @@ std::vector<int> rawparameter::getopnums(std::vector<int> disjregs)
     synchronize();
     
     std::vector<int> output(disjregs.size());
-    for (int i = 0; i < disjregs.size(); i++)
+    for (size_t i = 0; i < disjregs.size(); i++)
         output[i] = opnums[disjregs[i]];
     return output;
 }
@@ -75,7 +75,7 @@ void rawparameter::set(int physreg, expression input)
     // Consider ALL disjoint regions in the physical region with (-1):
     std::vector<int> selecteddisjregs = ((universe::mymesh->getphysicalregions())->get(physreg))->getdisjointregions(-1);
 
-    for (int i = 0; i < selecteddisjregs.size(); i++)
+    for (size_t i = 0; i < selecteddisjregs.size(); i++)
     {
         opnums[selecteddisjregs[i]] = maxopnum;
         for (int row = 0; row < mynumrows; row++)
@@ -150,14 +150,14 @@ std::vector<std::vector<densematrix>> rawparameter::interpolate(int row, int col
         // Preallocate the harmonics not yet in 'out':
         if (out.size() < currentinterp.size())
             out.resize(currentinterp.size());
-        for (int h = 0; h < currentinterp.size(); h++)
+        for (size_t h = 0; h < currentinterp.size(); h++)
         {
             if (currentinterp[h].size() == 1 && out[h].size() == 0)
                 out[h] = {densematrix(elemselect.countincurrentorientation(), evaluationcoordinates.size()/3, 0)};
         }
         
         // Insert 'currentinterp' in 'out':
-        for (int h = 0; h < currentinterp.size(); h++)
+        for (size_t h = 0; h < currentinterp.size(); h++)
         {
             if (currentinterp[h].size() == 1)
                 out[h][0].insertatrows(elemselect.getelementindexes(), currentinterp[h][0]);
@@ -206,9 +206,9 @@ densematrix rawparameter::multiharmonicinterpolate(int row, int col, int numtime
         
         std::vector<int> selectedelementindexes = elemselect.getelementindexes();
         std::vector<int> selectedcolumns(selectedelementindexes.size()*evaluationcoordinates.size()/3);
-        for (int j = 0; j < selectedelementindexes.size(); j++)
+        for (size_t j = 0; j < selectedelementindexes.size(); j++)
         {
-            for (int k = 0; k < evaluationcoordinates.size()/3; k++)
+            for (size_t k = 0; k < evaluationcoordinates.size()/3; k++)
                 selectedcolumns[j*evaluationcoordinates.size()/3+k] = selectedelementindexes[j]*evaluationcoordinates.size()/3+k;
         }
         out.insertatcolumns(selectedcolumns, myoperations[mydisjregs[0]][row*mynumcols+col]->multiharmonicinterpolate(numtimeevals, myselection, evaluationcoordinates, meshdeform));
@@ -239,7 +239,7 @@ void rawparameter::print(void)
     std::cout << std::endl;
     std::cout << "Printing parameter of size " << mynumrows << "x" << mynumcols;
     std::cout << std::endl;
-    for (int i = 0; i < myoperations.size(); i++)
+    for (size_t i = 0; i < myoperations.size(); i++)
     {   
         if (myoperations[i][0] != NULL)
         {
